refactor: simplify sortable list
